:recycle: Refactoring al mandar un grupo de codigo a un elemento de nivel inferior, era innecesario que se resuelva en un nivel superior, unicamente agregaba parametros
